refactor(controller): refactor controllers
refactor all the controllers
