+++ src/main/java/com/revature/controller/HotelRoomTypesController.java
@@ -18,6 +18,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+@CrossOrigin(origins = "http://localhost:5173", allowCredentials = "true")
 @RestController
 @RequestMapping("hotels/{hotelId}/roomTypes")
 public class HotelRoomTypesController {
